Move language specific code into shared location now that we have ICompilationUnitSyntax.
