+++ src/Workspaces/CSharp/Portable/Extensions/SyntaxTreeExtensions.cs
@@ -318,7 +318,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
         public static bool IsEntirelyWithinMultiLineComment(
             this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
         {
-            var trivia = FindTriviaAndAdjustForEndOfFile(syntaxTree, position, cancellationToken);
+            var trivia = syntaxTree.FindTriviaAndAdjustForEndOfFile(position, cancellationToken);
 
             if (trivia.IsMultiLineComment())
             {
@@ -335,7 +335,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
         public static bool IsEntirelyWithinTopLevelSingleLineComment(
             this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
         {
-            var trivia = FindTriviaAndAdjustForEndOfFile(syntaxTree, position, cancellationToken);
+            var trivia = syntaxTree.FindTriviaAndAdjustForEndOfFile(position, cancellationToken);
 
             if (trivia.Kind() == SyntaxKind.EndOfLineTrivia)
             {
@@ -361,7 +361,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
         {
             // Search inside trivia for directives to ensure that we recognize
             // single-line comments at the end of preprocessor directives.
-            var trivia = FindTriviaAndAdjustForEndOfFile(syntaxTree, position, cancellationToken, findInsideTrivia: true);
+            var trivia = syntaxTree.FindTriviaAndAdjustForEndOfFile(position, cancellationToken, findInsideTrivia: true);
 
             if (trivia.Kind() == SyntaxKind.EndOfLineTrivia)
             {
@@ -380,35 +380,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             }
 
             return false;
-        }
-
-        private static SyntaxTrivia FindTriviaAndAdjustForEndOfFile(
-            SyntaxTree syntaxTree, int position, CancellationToken cancellationToken, bool findInsideTrivia = false)
-        {
-            var root = syntaxTree.GetRoot(cancellationToken) as CompilationUnitSyntax;
-            var trivia = root.FindTrivia(position, findInsideTrivia);
-
-            // If we ask right at the end of the file, we'll get back nothing.
-            // We handle that case specially for now, though SyntaxTree.FindTrivia should
-            // work at the end of a file.
-            if (position == root.FullWidth())
-            {
-                var endOfFileToken = root.EndOfFileToken;
-                if (endOfFileToken.HasLeadingTrivia)
-                {
-                    trivia = endOfFileToken.LeadingTrivia.Last();
-                }
-                else
-                {
-                    var token = endOfFileToken.GetPreviousToken(includeSkipped: true);
-                    if (token.HasTrailingTrivia)
-                    {
-                        trivia = token.TrailingTrivia.Last();
-                    }
-                }
-            }
-
-            return trivia;
         }
 
         private static bool AtEndOfIncompleteStringOrCharLiteral(SyntaxToken token, int position, char lastChar)
@@ -557,44 +528,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             }
 
             return false;
-        }
-
-        public static SyntaxToken FindTokenOrEndToken(
-            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            Contract.ThrowIfNull(syntaxTree);
-
-            var root = syntaxTree.GetRoot(cancellationToken) as CompilationUnitSyntax;
-            var result = root.FindToken(position, findInsideTrivia: true);
-            if (result.Kind() != SyntaxKind.None)
-            {
-                return result;
-            }
-
-            // Special cases.  See if we're actually at the end of a:
-            // a) doc comment
-            // b) pp directive
-            // c) file
-
-            var triviaList = root.EndOfFileToken.LeadingTrivia;
-            foreach (var trivia in triviaList.Reverse())
-            {
-                if (trivia.HasStructure)
-                {
-                    var token = trivia.GetStructure().GetLastToken(includeZeroWidth: true);
-                    if (token.Span.End == position)
-                    {
-                        return token;
-                    }
-                }
-            }
-
-            if (position == root.FullSpan.End)
-            {
-                return root.EndOfFileToken;
-            }
-
-            return default(SyntaxToken);
         }
 
         public static IList<MemberDeclarationSyntax> GetMembersInSpan(
