+++ src/Workspaces/Core/Portable/Shared/Extensions/SyntaxTreeExtensions.cs
@@ -134,5 +134,74 @@ namespace Microsoft.CodeAnalysis.Shared.Extensions
 
             return position <= firstToken.SpanStart;
         }
+
+        public static SyntaxToken FindTokenOrEndToken(
+            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
+        {
+            Contract.ThrowIfNull(syntaxTree);
+
+            var root = syntaxTree.GetRoot(cancellationToken);
+            var compilationUnit = root as ICompilationUnitSyntax;
+            var result = root.FindToken(position, findInsideTrivia: true);
+            if (result.RawKind != 0)
+            {
+                return result;
+            }
+
+            // Special cases.  See if we're actually at the end of a:
+            // a) doc comment
+            // b) pp directive
+            // c) file
+
+            var triviaList = compilationUnit.EndOfFileToken.LeadingTrivia;
+            foreach (var trivia in triviaList.Reverse())
+            {
+                if (trivia.HasStructure)
+                {
+                    var token = trivia.GetStructure().GetLastToken(includeZeroWidth: true);
+                    if (token.Span.End == position)
+                    {
+                        return token;
+                    }
+                }
+            }
+
+            if (position == root.FullSpan.End)
+            {
+                return compilationUnit.EndOfFileToken;
+            }
+
+            return default(SyntaxToken);
+        }
+
+        internal static SyntaxTrivia FindTriviaAndAdjustForEndOfFile(
+            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken, bool findInsideTrivia = false)
+        {
+            var root = syntaxTree.GetRoot(cancellationToken);
+            var compilationUnit = root as ICompilationUnitSyntax;
+            var trivia = root.FindTrivia(position, findInsideTrivia);
+
+            // If we ask right at the end of the file, we'll get back nothing.
+            // We handle that case specially for now, though SyntaxTree.FindTrivia should
+            // work at the end of a file.
+            if (position == root.FullWidth())
+            {
+                var endOfFileToken = compilationUnit.EndOfFileToken;
+                if (endOfFileToken.HasLeadingTrivia)
+                {
+                    trivia = endOfFileToken.LeadingTrivia.Last();
+                }
+                else
+                {
+                    var token = endOfFileToken.GetPreviousToken(includeSkipped: true);
+                    if (token.HasTrailingTrivia)
+                    {
+                        trivia = token.TrailingTrivia.Last();
+                    }
+                }
+            }
+
+            return trivia;
+        }
     }
 }
